test_requirements.txt
Django < 1.9
django-cms==3.2.0
icalendar
mock
